requirements.txt
black==23.3.0
Django==4.1.6
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flake8==6.0.0
isort==5.12.0
psycopg2-binary==2.9.9